Renamed Mongo ORM config file.
