+++ src/server/app/app.module.ts
@@ -3,7 +3,7 @@ import { AppController } from "./app.controller";
 import { AppService } from "./app.service";
 import { ConfigModule } from "@nestjs/config";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import * as connectionOptions from "../ormconfig";
+import * as connectionOptions from "../ormconfigMongo";
 
 @Module({
     imports: [
